pyproject.toml
[project]
name = "temporal-ai-agents-workshop"
version = "1.0.0"
description = "Workshop: Building Durable AI Agents with OpenAI + Temporal"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.0",
    "temporalio>=1.7.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "notebook>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
